Remove re-renders on board
